Only add .0 to numerics, not integers

+++ R/translate-sql.r
@@ -79,7 +79,7 @@ sql_unary <- function(f) {
 }
 sql_atomic <- function(x) {
   if (is.character(x)) x <- paste0('"', x, '"')
-  if (is.numeric(x) && any(is.wholenumber(x))) {
+  if (is.double(x) && any(is.wholenumber(x))) {
     x <- sprintf("%.1f", x)
   }
